refactor(product page): split server code from client code
